Migrate from PostgreSQL to SQLite
- As part of the ongoing efforts to reduce the number of dependencies
  sms-irc requires, this commit completely removes PostgreSQL support,
  replacing it instead with SQLite through the `sequelight` crate
  (recently extracted from trainsplorer).
- While we were at it, handling of PduAddresses and Jids was also
  improved; now, these types are represented as-is in models.rs,
  and are transparently serialized/deserialized when interacting with
  the database, removing the need for hacky code that does this all
  about the place.
- No migration path is provided yet; the intent is to provide this in
  a separate crate.

+++ src/sender_common.rs
@@ -22,7 +22,7 @@ pub trait Sender {
     fn ensure_joined(&mut self, _ch: &str) -> Result<()> {
         Ok(())
     }
-    fn send_raw_message(&mut self, from_nick: &str, msg: &str, group_target: Option<i32>) -> Result<()> {
+    fn send_raw_message(&mut self, from_nick: &str, msg: &str, group_target: Option<i64>) -> Result<()> {
         let dest = if let Some(g) = group_target {
             let grp = self.store().get_group_by_id(g)?;
             self.ensure_joined(&grp.channel)?;
